Added the eeprom interface fully to the design

+++ include/EEPROMInterface.h
@@ -58,14 +58,14 @@ public:
     }
 
     static uint16_t read(uint8_t targetPage, uint8_t offset, LoadStoreStyle) noexcept {
-        SplitWord16 address{offset, targetPage - StartPage}; // normalize it
+        SplitWord16 address{offset, static_cast<byte>(targetPage - StartPage)}; // normalize it
         uint16_t result = 0;
         EEPROM.get(address.getWholeValue(), result);
         return result;
     }
 
     static void write(uint8_t targetPage, uint8_t offset, LoadStoreStyle lss, SplitWord16 value) noexcept {
-        SplitWord16 address{offset, targetPage - StartPage};
+        SplitWord16 address{offset, static_cast<byte>(targetPage - StartPage)};
         switch (lss) {
             case LoadStoreStyle::Lower8:
                 EEPROM.update(address.getWholeValue(), value.bytes[0]);
